requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pillow>=8.0.0
matplotlib>=3.3.4
scikit-learn>=0.24.2
pandas>=1.3.0
tqdm>=4.61.1
opencv-python>=4.5.3